Add content to the breakfast omelette html file
This commit adds contents to the breakfast-omelette.html file.
Before this commit the html file was empty.
This meant that no content was displayed upon displaying the html file
inside a browser.
Now the breakfast-omelette.html file has been filled with content and
displays the recipe for a breakfast omelette. It accomplishes this
using the following parts in the following order:
 - a heading containing the name of the recipe: breakfast omelette
 - an image of an omelette which is stored in the images directory of
   the project
 - a description of the recipe under the a description heading
 - an ingredients list under the ingredients heading
 - an ordered list of steps to make the omelette under the steps
   heading