Adding new tab link targets to the recipe links
This commit adds the target attribute to the anchor elements in the
index.html file. The value assigned to the target attributes is _blank
which is a predefined value for the target attribute that carries
special meaning for web browsers. Specifically, the "_blank" value
tells the browser to open a new tab when the anchor element (link)
containing this target value is pressed.