Add external CSS to index.html
This commit creates a CSS file in the odin recipes repository and links
the homepage index.html file to the external CSS file.
Some type selectors, an id selector, and a grouping selector were added
to this css file to style the elements of the index.html file.
An animated GIF of Gordon Ramsay is added below the heading and some
colours are added to the writing on the page.