Add CSS classes and external CSS link to recipes
This commit adds an external css link from the recipe html pages to the
style.css file.

This commit also adds classes to the elements of the recipe html files:

 * recipe-name to the first heading 1 element which contains the recipe
   name

 * dish-image to the image element

 * description to the heading 2 element called description

 * ingredients to the heading 2 element called ingredients

 * steps to the heading 2 element called steps

This commit adds the classes added to the html files to the style.css
file in the form of class selectors for CSS rule sets yet to be
applied.